+++ packages/docs/docs/main/change-log.md
@@ -61,6 +61,8 @@ This page lists the changes to the IsaacScript framework.
   - `CHEST_PICKUP_VARIANTS_SET`
   - `COLLECTIBLE_NAME_TO_TYPE_MAP`
   - `TRINKET_NAME_TO_TYPE_MAP`
+- Added the following helper types:
+  - `CompositionTypeSatisfiesEnum`
 - Renamed the following features:
   - `ISCFeature.PLAYER_INVENTORY` --> `ISCFeature.PLAYER_COLLECTIBLE_TRACKING`
   - `ISCFeature.EDEN_STARTING_STATS` --> `ISCFeature.EDEN_STARTING_STATS_HEALTH`